requirements.txt
pyrogram>=2.0.100,<3.0.0
TgCrypto>=1.2.5
yt-dlp>=2024.05.27
python-dotenv>=1.0.0
uvloop>=0.19.0; platform_system != 'Windows'
